refactor(encrypted-archive): move the `getCryptoAlgorithm()` function from `src/index.node.ts` into `src/node/cipher.ts`
