+++ packages/encrypted-archive/src/node/utils/index.ts
@@ -1,5 +1,8 @@
 import { inspect as nodeInspect } from 'util';
 
+import { isPropAccessible } from '@sounisi5011/ts-utils-is-property-accessible';
+
+import { isString } from '../../core/utils/index';
 import type { InspectFn } from '../../types/builtin';
 
 export const inspect: InspectFn = value => nodeInspect(value, { breakLength: Infinity });
@@ -19,4 +22,65 @@ export function bufferFrom(
     if (typeof value === 'string') return Buffer.from(value, encoding);
     if (ArrayBuffer.isView(value)) return arrayBufferView2NodeBuffer(value);
     return Buffer.from(value);
+}
+
+function isErrorConstructor(value: unknown): value is ErrorConstructor {
+    /**
+     * @see https://stackoverflow.com/a/14486171/4907315
+     */
+    return typeof value === 'function' && value.prototype instanceof Error;
+}
+
+export function fixNodePrimordialsErrorInstance(oldError: unknown): never {
+    if (
+        typeof oldError === 'object' && isPropAccessible(oldError)
+        && !(oldError instanceof Error)
+        && typeof oldError['name'] === 'string'
+        && typeof oldError['message'] === 'string'
+    ) {
+        const name = oldError['name'];
+        const detectConstructor = (globalThis as Record<string, unknown>)[name];
+        const ErrorConstructor = isErrorConstructor(detectConstructor) ? detectConstructor : Error;
+        const newError = new ErrorConstructor();
+        Object.defineProperties(
+            newError,
+            Object.getOwnPropertyDescriptors(oldError),
+        );
+        throw newError;
+    }
+    throw oldError;
+}
+
+/**
+ * The Node.js built-in function throws a pseudo Error object that does not extend the Error object.
+ * Their stack trace is incomplete and errors are difficult to trace.
+ * For example, the results of {@link https://jestjs.io/ Jest} do not show where the error occurred.
+ * This function will fix the stack trace of the pseudo Error object to the proper one.
+ */
+export function fixNodePrimordialsErrorStackTrace(oldError: unknown): never {
+    /* eslint-disable @typescript-eslint/no-throw-literal */
+    if (!(typeof oldError === 'object' && isPropAccessible(oldError) && !(oldError instanceof Error))) throw oldError;
+    const oldErrorStackTrace = oldError['stack'];
+
+    if (!(isString(oldError['message']) && isString(oldErrorStackTrace))) throw oldError;
+    const newError = new Error();
+    const newErrorStackTrace = newError.stack;
+
+    if (!isString(newErrorStackTrace)) throw oldError;
+    const newErrorStackTraceLastLine = newErrorStackTrace.replace(/^(?:(?!\n[^\n]+$).)+/s, '');
+
+    if (oldErrorStackTrace.includes(newErrorStackTraceLastLine)) throw oldError;
+    Object.defineProperties(
+        newError,
+        Object.getOwnPropertyDescriptors(oldError),
+    );
+    newError.stack = oldErrorStackTrace
+        + newErrorStackTrace.replace(
+            new RegExp(
+                String.raw`^Error: [^\n]*(?:\n *at (?:Object\.)?${fixNodePrimordialsErrorStackTrace.name}\b[^\n]*)?`,
+            ),
+            '',
+        );
+    throw newError;
+    /* eslint-enable */
 }
