+++ packages/encrypted-archive/tests/unit/cipher.ts
@@ -1,97 +1,64 @@
 import { randomBytes } from 'crypto';
 
-import { cryptoAlgorithmMap } from '../../src/node/cipher';
+import type { CryptoAlgorithmName } from '../../src/core/types/crypto';
+import { getCryptoAlgorithm } from '../../src/node/cipher';
 
-describe.each([...cryptoAlgorithmMap.values()].map(algorithm => [algorithm.name, algorithm] as const))(
-    '%s',
+const algorithmNameList = ['aes-256-gcm', 'chacha20-poly1305'] as const;
+const cases = algorithmNameList
+    .map(algorithmName => [algorithmName, getCryptoAlgorithm(algorithmName)] as const)
+    .filter(<K, V>(value: readonly [K, V | undefined]): value is [K, V] => value[1] !== undefined);
+
+describe.each(cases)(
+    'getCryptoAlgorithm(algorithmName: %o)',
     (_, algorithm) => {
         const key = randomBytes(algorithm.keyLength);
         const nonce = randomBytes(algorithm.nonceLength);
         const cleartext = randomBytes(42);
 
-        it('createCipher()', async () => {
-            let cipher: ReturnType<typeof algorithm.createCipher>;
-            expect(() => {
-                cipher = algorithm.createCipher(key, nonce);
-            }).not.toThrow();
-            expect(() => cipher.update(cleartext)).not.toThrow();
-            expect(() => cipher.final()).not.toThrow();
+        it('encrypt()', async () => {
+            const result = algorithm.encrypt({ key, nonce, cleartext });
+            await expect(result).toResolve();
         });
 
-        it('createDecipher()', async () => {
-            const cipher = algorithm.createCipher(key, nonce);
-            const ciphertext = Buffer.concat([cipher.update(cleartext), cipher.final()]);
-            const authTag = cipher.getAuthTag();
-
-            let decipher: ReturnType<typeof algorithm.createDecipher>;
-            expect(() => {
-                decipher = algorithm.createDecipher(key, nonce);
-            }).not.toThrow();
-            expect(() => decipher.setAuthTag(authTag)).not.toThrow();
-            expect(() => decipher.update(ciphertext)).not.toThrow();
-            expect(() => decipher.final()).not.toThrow();
+        it('decrypt()', async () => {
+            const { ciphertext, authTag } = await algorithm.encrypt({ key, nonce, cleartext });
+            const result = algorithm.decrypt({ key, nonce, authTag, ciphertext });
+            await expect(result).toResolve();
         });
 
         it('match cleartext and ciphertext', async () => {
-            const cipher = algorithm.createCipher(key, nonce);
-            const ciphertext = Buffer.concat([cipher.update(cleartext), cipher.final()]);
-            const authTag = cipher.getAuthTag();
-
-            let decipher: ReturnType<typeof algorithm.createDecipher>;
-            expect(() => {
-                decipher = algorithm.createDecipher(key, nonce);
-            }).not.toThrow();
-
-            expect(() => decipher.setAuthTag(authTag)).not.toThrow();
-
-            const cleartextPartList: Buffer[] = [];
-
-            expect(() => cleartextPartList.push(decipher.update(ciphertext))).not.toThrow();
-
-            expect(() => cleartextPartList.push(decipher.final())).not.toThrow();
-
-            const cleartext2 = Buffer.concat(cleartextPartList);
+            const { ciphertext, authTag } = await algorithm.encrypt({ key, nonce, cleartext });
+            const { cleartext: cleartext2 } = await algorithm.decrypt({ key, nonce, authTag, ciphertext });
             expect(cleartext2).toBytesEqual(cleartext);
-            expect(cleartext).toBytesEqual(cleartext2);
         });
 
         describe('decryption fail', () => {
             it('different key', async () => {
-                const cipher = algorithm.createCipher(key, nonce);
-                const ciphertext = Buffer.concat([cipher.update(cleartext), cipher.final()]);
-                const authTag = cipher.getAuthTag();
-
+                const { ciphertext, authTag } = await algorithm.encrypt({ key, nonce, cleartext });
                 const key2 = randomBytes(algorithm.keyLength);
-
-                let decipher: ReturnType<typeof algorithm.createDecipher>;
-                expect(() => {
-                    decipher = algorithm.createDecipher(key2, nonce);
-                }).not.toThrow();
-
-                expect(() => decipher.setAuthTag(authTag)).not.toThrow();
-
-                expect(() => decipher.update(ciphertext)).not.toThrow();
-
-                expect(() => decipher.final()).toThrow(/^Unsupported state or unable to authenticate data$/);
+                const result = algorithm.decrypt({ key: key2, nonce, authTag, ciphertext });
+                await expect(result).rejects.toThrowWithMessage(
+                    Error,
+                    'Unsupported state or unable to authenticate data',
+                );
             });
             it('different nonce', async () => {
-                const cipher = algorithm.createCipher(key, nonce);
-                const ciphertext = Buffer.concat([cipher.update(cleartext), cipher.final()]);
-                const authTag = cipher.getAuthTag();
-
+                const { ciphertext, authTag } = await algorithm.encrypt({ key, nonce, cleartext });
                 const nonce2 = randomBytes(algorithm.nonceLength);
-
-                let decipher: ReturnType<typeof algorithm.createDecipher>;
-                expect(() => {
-                    decipher = algorithm.createDecipher(key, nonce2);
-                }).not.toThrow();
-
-                expect(() => decipher.setAuthTag(authTag)).not.toThrow();
-
-                expect(() => decipher.update(ciphertext)).not.toThrow();
-
-                expect(() => decipher.final()).toThrow(/^Unsupported state or unable to authenticate data$/);
+                const result = algorithm.decrypt({ key, nonce: nonce2, authTag, ciphertext });
+                await expect(result).rejects.toThrowWithMessage(
+                    Error,
+                    'Unsupported state or unable to authenticate data',
+                );
             });
         });
     },
 );
+
+describe('unknown algorithm', () => {
+    it('should return `undefined`', () => {
+        // @ts-expect-error TS2322: Type '"foo"' is not assignable to type 'CryptoAlgorithmName'.
+        const algorithmName: CryptoAlgorithmName = 'foo';
+        expect(getCryptoAlgorithm(algorithmName)).toBeUndefined();
+    });
+});
